Move queuing logic from orchestrator to activity
Jira ticket: CAMS-461
