+++ backend/function-apps/migration/orchestration/sub-orchestrator-queue-etl.ts
@@ -1,43 +1,21 @@
-import { output } from '@azure/functions';
 import {
   AcmsPredicateAndPage,
   AcmsPageReport,
-  AcmsEtlQueueItem,
 } from '../../../lib/use-cases/acms-orders/acms-orders';
 import { GET_CONSOLIDATIONS } from '../loadConsolidations';
 import { OrchestrationContext } from 'durable-functions';
-
-import * as dotenv from 'dotenv';
-dotenv.config();
-
-const etlQueueOutput = output.storageQueue({
-  queueName: process.env.CAMS_MIGRATION_TASK_QUEUE,
-  connection: 'AzureWebJobs',
-});
 
 export function* subOrchestratorETL(context: OrchestrationContext) {
   const predicateAndPage: AcmsPredicateAndPage = context.df.getInput();
 
   const leadCaseIds = yield context.df.callActivity(GET_CONSOLIDATIONS, predicateAndPage);
 
-  const queueItems = [];
-  for (let i = 0; i < leadCaseIds.length; i++) {
-    const leadCaseIdString = leadCaseIds[i].toString();
-    const queueItem: AcmsEtlQueueItem = {
-      divisionCode: predicateAndPage.divisionCode,
-      chapter: predicateAndPage.chapter,
-      leadCaseId: leadCaseIdString,
-    };
-    queueItems.push(queueItem);
-  }
-  context.extraOutputs.set(etlQueueOutput, queueItems);
-
   // TODO: Need to reconsider the report now that we are queuing the ETL tasks. These sane defaults are here to keep the rest of the orchestration from breaking.
   const result: AcmsPageReport = {
     predicateAndPage,
     successful: {
       childCaseCount: 0,
-      leadCaseCount: queueItems.length,
+      leadCaseCount: leadCaseIds.length,
     },
     failed: {
       childCaseCount: 0,
